Title: 특정 구간의 원소 평균값, Time: 8ms, Memory: 0MB, Status: Passed - Codetree

+++ 260220/특정 구간의 원소 평균값/elemental-mean-value-for-a-particular-interval.cpp
@@ -6,7 +6,7 @@ int main() {
     //개념 정리. 3중 포문. i는 0 시작 n 끝. j는 i 시작 n 끝. i는 시작점, j는 도착점을 의미. k는 i 이상 j 미만. 즉 구간.
     int n;
     cin >> n;
-    float arr[n] = {};
+    int arr[n] = {};
     
     for(int i = 0; i < n; i++) {
         cin >> arr[i];
@@ -17,7 +17,7 @@ int main() {
         for(int j = i; j < n; j++) {
             bool exist = false;
             float avg = 0;
-            float count = 0;
+            int count = 0;
             for(int k = i; k <= j; k++) {
                 //먼저 평균을 구해야한다. j에서 i를 빼면? 개수다. 그걸로 일단 먼저 순회하고 평균 구하고 평균 있나 보자. 4중 for문 될듯.
                 avg += arr[k];
